requirements.txt
streamlit
googletrans==4.0.0-rc1
gTTS
SpeechRecognition
pytesseract
PyPDF2
python-docx
Pillow
sounddevice
numpy
